+++ src/app/app-routing.module.ts
@@ -5,11 +5,12 @@ import { LayoutComponent } from './shared/components/layout/layout.component';
 
 const routes: Routes = [
   {
-    path: 'guides',
+    path: 'guides/:partnerId',
     component: LayoutComponent,
     loadChildren: () => import('./pages/pages.module').then(m => m.PagesModule)
   },
-  { path: '', pathMatch: 'full', redirectTo: '/guides' },
+  { path: 'guides', redirectTo: '/guides/0' },
+  { path: '', pathMatch: 'full', redirectTo: '/guides/0' },
   { path: '**', component: NotFoundPageComponent }
 ];