+++ src/app/shared/components/layout/layout.component.ts
@@ -1,11 +1,15 @@
 import { Component, OnDestroy, OnInit } from "@angular/core";
-import { Data } from "@angular/router";
+import { ActivatedRoute, Data } from "@angular/router";
 import { select, Store } from "@ngrx/store";
 import { Subject } from "rxjs";
 import { takeUntil } from "rxjs/operators";
 import { environment } from "src/environments/environment";
 import { AppState } from "../../ngrx";
+
 import { selectRouteData } from "../../ngrx/router/router.actions";
+import { getCurrentPartnerId } from "../../ngrx/global/global.selectors";
+import * as globalActions from "../../ngrx/global/global.actions";
+
 
 @Component({
   selector: "app-layout",
@@ -20,16 +24,15 @@ export class LayoutComponent implements OnInit, OnDestroy {
   private _unsubcriber: Subject<any> = new Subject<any>();
 
   constructor(
-    private _store: Store<AppState>
+    private _store: Store<AppState>,
+    private _route: ActivatedRoute,
   ) {
-    this.components.push({ name: 'CFAR Offer Banner', link: 'cfar-offer-banner' });
-    this.components.push({ name: 'CFAR Offer Banner Large', link: 'cfar-offer-banner-large' });
-    this.components.push({ name: 'CFAR Offer Dialog', link: 'cfar-offer-dialog' });
-    this.components.push({ name: 'CFAR Exercise Flow', link: 'cfar-exercise-flow' });
-    
     // Working items
     if (!environment.production) {
     }
+
+    // Set the partnerId in the state
+    this._store.dispatch(globalActions.setCurrentPartnerId({ currentPartnerId: this._route.snapshot.params.partnerId }));
   }
 
   // -------------------------------
@@ -42,6 +45,12 @@ export class LayoutComponent implements OnInit, OnDestroy {
       select(selectRouteData),
       takeUntil(this._unsubcriber)
     ).subscribe((data: Data) => this.selectedPath = data.path);
+
+    // Update the visibles components when the partnerId has changed
+    this._store.pipe(
+      select(getCurrentPartnerId),
+      takeUntil(this._unsubcriber)
+    ).subscribe((partnerId: number) => this._loadPartnerData(partnerId));
   }
 
   ngOnDestroy(): void {
@@ -51,5 +60,34 @@ export class LayoutComponent implements OnInit, OnDestroy {
 
   onScrollToTop(event: any): void {
     document.querySelector('#scroll-anchor')?.scrollTo(0,0);
+  }
+
+  // -------------------------------
+  // - PRIVATES METHODS
+  // -------------------------------
+
+  /**
+   * FIXME : Load config by partner ?
+   * @param partnerId 
+   */
+  private _loadPartnerData(partnerId: number): void {
+    var components = [];
+
+    if (partnerId == 0) {
+      components.push({ name: 'CFAR Offer Banner', link: 'cfar-offer-banner' });
+      components.push({ name: 'CFAR Offer Banner Large', link: 'cfar-offer-banner-large' });
+      components.push({ name: 'CFAR Offer Dialog', link: 'cfar-offer-dialog' });
+      components.push({ name: 'CFAR Exercise Flow', link: 'cfar-exercise-flow' });
+    } else if (partnerId == 1) {
+      components.push({ name: 'CFAR Offer Dialog', link: 'cfar-offer-dialog' });
+      components.push({ name: 'CFAR Exercise Flow', link: 'cfar-exercise-flow' });
+    } else {
+      components.push({ name: 'CFAR Offer Banner', link: 'cfar-offer-banner' });
+      components.push({ name: 'CFAR Offer Banner Large', link: 'cfar-offer-banner-large' });
+      components.push({ name: 'CFAR Offer Dialog', link: 'cfar-offer-dialog' });
+      components.push({ name: 'CFAR Exercise Flow', link: 'cfar-exercise-flow' });
+    }
+
+    this.components = components;
   }
 }