+++ src/app/shared/enums/theme.enum.ts
@@ -1,0 +1,6 @@
+export enum Theme {
+    THEME_DARK_HOPPER = "theme-dark-hopper",
+    THEME_LIGHT_HOPPER = "theme-light-hopper",
+    THEME_LIGHT_BLUE = "theme-light-blue",
+    THEME_LIGHT_YELLOW = "theme-light-yellow"
+}